+++ src/urllib3/_async/connectionpool.py
@@ -436,6 +436,10 @@ class AsyncHTTPConnectionPool(AsyncConnectionPool, AsyncRequestMethods):
     def is_idle(self) -> bool:
         return self.pool is None or self.pool.bag_only_idle
 
+    @property
+    def is_saturated(self) -> bool:
+        return self.pool is not None and self.pool.bag_only_saturated
+
     async def _new_conn(
         self, *, heb_timeout: Timeout | None = None
     ) -> AsyncHTTPConnection: