+++ src/urllib3/backend/_async/hface.py
@@ -131,8 +131,18 @@ class AsyncHfaceBackend(AsyncBaseBackend):
     @property
     def is_saturated(self) -> bool:
         if self._protocol is None:
-            return True
-        return self._protocol.is_available() is False
+            return False
+        # is_available also includes whether we must goaway
+        # we want to focus on stream capacity here.
+        if self._protocol.is_available() is False:
+            return self._protocol.has_expired() is False
+        return False
+
+    @property
+    def max_stream_count(self) -> int:
+        if self._protocol is None:
+            return 0
+        return self._protocol.max_stream_count
 
     @property
     def is_multiplexed(self) -> bool:
@@ -1304,7 +1314,7 @@ class AsyncHfaceBackend(AsyncBaseBackend):
                 pass  # Hmm... this should be impossible.
 
             # remote can refuse future inquiries, so no need to go further with this conn.
-            if self._protocol.has_expired():  # type: ignore[union-attr]
+            if self._protocol.is_idle() and self._protocol.has_expired():  # type: ignore[union-attr]
                 await self.close()
             return b"", True, None
 
@@ -1334,7 +1344,7 @@ class AsyncHfaceBackend(AsyncBaseBackend):
                 pass  # Hmm... this should be impossible.
 
             # remote can refuse future inquiries, so no need to go further with this conn.
-            if self._protocol.has_expired():  # type: ignore[union-attr]
+            if self._protocol.is_idle() and self._protocol.has_expired():  # type: ignore[union-attr]
                 await self.close()
             elif self.is_idle:
                 # probe for h3/quic if available, and remember it.