+++ src/urllib3/contrib/hface/protocols/http2/_h2.py
@@ -21,6 +21,7 @@ from typing import Iterator
 
 import jh2.config  # type: ignore
 import jh2.connection  # type: ignore
+import jh2.errors  # type: ignore
 import jh2.events  # type: ignore
 import jh2.exceptions  # type: ignore
 import jh2.settings  # type: ignore
@@ -70,6 +71,43 @@ class _PatchedH2Connection(jh2.connection.H2Connection):  # type: ignore[misc]
             return self._observable_impl._open_stream_count
         return super()._open_streams(*args, **kwargs)  # type: ignore[no-any-return]
 
+    def _receive_goaway_frame(self, frame):  # type: ignore[no-untyped-def]
+        """
+        Receive a GOAWAY frame on the connection.
+        We purposely override this method to work around a known bug of jh2.
+        """
+        events = self.state_machine.process_input(
+            jh2.connection.ConnectionInputs.RECV_GOAWAY
+        )
+
+        err_code = jh2.errors._error_code_from_int(frame.error_code)
+
+        # GOAWAY allows an
+        # endpoint to gracefully stop accepting new streams while still
+        # finishing processing of previously established streams.
+        # see https://tools.ietf.org/html/rfc7540#section-6.8
+        # hyper/h2 does not allow such a thing for now. let's work around this.
+        if (
+            err_code == 0
+            and self._observable_impl is not None
+            and self._observable_impl._open_stream_count > 0
+        ):
+            self.state_machine.state = jh2.connection.ConnectionState.CLIENT_OPEN
+
+        # Clear the outbound data buffer: we cannot send further data now.
+        self.clear_outbound_data_buffer()
+
+        # Fire an appropriate ConnectionTerminated event.
+        new_event = jh2.events.ConnectionTerminated()
+        new_event.error_code = err_code
+        new_event.last_stream_id = frame.last_stream_id
+        new_event.additional_data = (
+            frame.additional_data if frame.additional_data else None
+        )
+        events.append(new_event)
+
+        return [], events
+
 
 class HTTP2ProtocolHyperImpl(HTTP2Protocol):
     implementation: str = "h2"
@@ -97,20 +135,26 @@ class HTTP2ProtocolHyperImpl(HTTP2Protocol):
         self._events: StreamMatrix = StreamMatrix()
         self._terminated: bool = False
         self._goaway_to_honor: bool = False
-        self._max_stream_count = self._connection.remote_settings.max_concurrent_streams
-        self._max_frame_size = self._connection.remote_settings.max_frame_size
+        self._max_stream_count: int = (
+            self._connection.remote_settings.max_concurrent_streams
+        )
+        self._max_frame_size: int = self._connection.remote_settings.max_frame_size
 
     def max_frame_size(self) -> int:
-        return self._max_frame_size  # type: ignore[no-any-return]
+        return self._max_frame_size
 
     @staticmethod
     def exceptions() -> tuple[type[BaseException], ...]:
         return jh2.exceptions.ProtocolError, jh2.exceptions.H2Error
 
     def is_available(self) -> bool:
-        if self._terminated or self._goaway_to_honor:
+        if self._terminated:
             return False
-        return self._max_stream_count > self._open_stream_count  # type: ignore[no-any-return]
+        return self._max_stream_count > self._open_stream_count
+
+    @property
+    def max_stream_count(self) -> int:
+        return self._max_stream_count
 
     def is_idle(self) -> bool:
         return self._terminated is False and self._open_stream_count == 0
@@ -196,7 +240,10 @@ class HTTP2ProtocolHyperImpl(HTTP2Protocol):
                 # so we emit an event called "GoawayReceived".
                 if e.error_code == 0:
                     self._goaway_to_honor = True
-                yield GoawayReceived(e.last_stream_id, e.error_code)
+                    yield GoawayReceived(e.last_stream_id, e.error_code)
+                else:
+                    self._terminated = True
+                    yield ConnectionTerminated(e.error_code, None)
             elif isinstance(e, jh2.events.SettingsAcknowledged):
                 yield HandshakeCompleted(alpn_protocol="h2")
 
@@ -210,16 +257,6 @@ class HTTP2ProtocolHyperImpl(HTTP2Protocol):
         if not data:
             return
 
-        # GOAWAY allows an
-        # endpoint to gracefully stop accepting new streams while still
-        # finishing processing of previously established streams.
-        # see https://tools.ietf.org/html/rfc7540#section-6.8
-        # hyper/h2 does not allow such a thing for now. let's work around this.
-        if self._goaway_to_honor and not self._terminated and self._open_stream_count:
-            self._connection.state_machine.state = (
-                jh2.connection.ConnectionState.CLIENT_OPEN
-            )
-
         try:
             h2_events = self._connection.receive_data(data)
         except jh2.exceptions.ProtocolError as e:
@@ -227,10 +264,15 @@ class HTTP2ProtocolHyperImpl(HTTP2Protocol):
         else:
             self._events.extend(self._map_events(h2_events))
 
+        # we want to perpetually mark the connection as "saturated"
+        if self._goaway_to_honor:
+            self._max_stream_count = self._open_stream_count
+
         if self._connection.remote_settings.has_update:
-            self._max_stream_count = (
-                self._connection.remote_settings.max_concurrent_streams
-            )
+            if not self._goaway_to_honor:
+                self._max_stream_count = (
+                    self._connection.remote_settings.max_concurrent_streams
+                )
             self._max_frame_size = self._connection.remote_settings.max_frame_size
 
     def bytes_to_send(self) -> bytes: